+++ StarLevelSystem/modules/LevelSystem.cs
@@ -339,31 +339,31 @@ namespace StarLevelSystem.modules
             yield break;
         }
 
-        [HarmonyPatch(typeof(Fish), nameof(Fish.Awake))]
-        public static class RandomFishLevelExtension
-        {
-            public static void Postfix(Fish __instance) {
-                if (ValConfig.EnableScalingFish.Value == false) { return; }
-                if (__instance.m_nview == null || __instance.m_nview.GetZDO() == null) {  return; }
-                int storedLevel = __instance.m_nview.GetZDO().GetInt("SLE_Fish", 0);
-                if (storedLevel == 0) {
-                    LevelSystem.SelectCreatureBiomeSettings(__instance.gameObject, out string creature_name, out DataObjects.CreatureSpecificSetting creature_settings, out BiomeSpecificSetting biome_settings, out Heightmap.Biome biome);
-                    storedLevel = LevelSystem.DetermineLevel(__instance.gameObject, creature_name, creature_settings, biome_settings, ValConfig.FishMaxLevel.Value);
-                    __instance.m_nview.GetZDO().Set("SLE_Fish", storedLevel);
-                    __instance.GetComponent<ItemDrop>().SetQuality(storedLevel);
-                    __instance.GetComponent<ItemDrop>().m_itemData.m_shared.m_maxQuality = storedLevel + 1;
-                }
-                if (storedLevel > 1) {
-                    float scale = 1 + (ValConfig.FishSizeScalePerLevel.Value * storedLevel);
-                    __instance.GetComponent<ItemDrop>().SetQuality(storedLevel);
-                    __instance.GetComponent<ItemDrop>().m_itemData.m_shared.m_maxQuality = storedLevel + 1;
-                    //Logger.LogDebug($"Setting Fish level {storedLevel} size {scale}.");
-                    __instance.transform.localScale *= scale;
-                    Physics.SyncTransforms();
+        //[HarmonyPatch(typeof(Fish), nameof(Fish.Awake))]
+        //public static class RandomFishLevelExtension
+        //{
+        //    public static void Postfix(Fish __instance) {
+        //        if (ValConfig.EnableScalingFish.Value == false) { return; }
+        //        if (__instance.m_nview == null || __instance.m_nview.GetZDO() == null) {  return; }
+        //        int storedLevel = __instance.m_nview.GetZDO().GetInt("SLE_Fish", 0);
+        //        if (storedLevel == 0) {
+        //            LevelSystem.SelectCreatureBiomeSettings(__instance.gameObject, out string creature_name, out DataObjects.CreatureSpecificSetting creature_settings, out BiomeSpecificSetting biome_settings, out Heightmap.Biome biome);
+        //            storedLevel = LevelSystem.DetermineLevel(__instance.gameObject, creature_name, creature_settings, biome_settings, ValConfig.FishMaxLevel.Value);
+        //            __instance.m_nview.GetZDO().Set("SLE_Fish", storedLevel);
+        //            __instance.GetComponent<ItemDrop>().SetQuality(storedLevel);
+        //            __instance.GetComponent<ItemDrop>().m_itemData.m_shared.m_maxQuality = storedLevel + 1;
+        //        }
+        //        if (storedLevel > 1) {
+        //            float scale = 1 + (ValConfig.FishSizeScalePerLevel.Value * storedLevel);
+        //            __instance.GetComponent<ItemDrop>().SetQuality(storedLevel);
+        //            __instance.GetComponent<ItemDrop>().m_itemData.m_shared.m_maxQuality = storedLevel + 1;
+        //            //Logger.LogDebug($"Setting Fish level {storedLevel} size {scale}.");
+        //            __instance.transform.localScale *= scale;
+        //            Physics.SyncTransforms();
 
-                }
-            }
-        }
+        //        }
+        //    }
+        //}
 
         public static void UpdateMaxLevel() {
             IEnumerable<GameObject> fishes = Resources.FindObjectsOfTypeAll<GameObject>().Where(obj => obj.name.StartsWith("Fish"));
@@ -375,6 +375,38 @@ namespace StarLevelSystem.modules
             }
         }
 
+        public static int DeterministicDetermineTreeLevel(GameObject go)
+        {
+            Vector3 p = go.transform.position;
+            float distance_from_center = Vector2.Distance(new Vector2(p.x, p.y), new Vector2(center.x, center.z));
+            return Mathf.RoundToInt(distance_from_center / (WorldGenerator.worldSize / ValConfig.TreeMaxLevel.Value));
+        }
+
+        public static IEnumerator ModifyTreeWithLevel(TreeBase tree, int level)
+        {
+            yield return new WaitForSeconds(1f);
+            if (tree == null) { yield break; }
+            //Logger.LogDebug($"Tree level set to: {level}");
+            float scale = 1 + (ValConfig.TreeSizeScalePerLevel.Value * level);
+            tree.m_health += (tree.m_health * 0.1f * level);
+            // Logger.LogDebug($"Setting Tree size {scale}.");
+            tree.transform.localScale *= scale;
+            List<DropTable.DropData> drops = new List<DropTable.DropData>();
+            foreach (var drop in tree.m_dropWhenDestroyed.m_drops)
+            {
+                DropTable.DropData lvlupdrop = new DropTable.DropData();
+                // Scale the amount of drops based on level
+                lvlupdrop.m_stackMin = Mathf.RoundToInt(drop.m_stackMin * (ValConfig.PerLevelLootScale.Value * level));
+                lvlupdrop.m_stackMax = Mathf.RoundToInt(drop.m_stackMax * (ValConfig.PerLevelLootScale.Value * level));
+                // Logger.LogDebug($"Scaling drop {drop.m_item.name} from {drop.m_stackMin}-{drop.m_stackMax} to {lvlupdrop.m_stackMin}-{lvlupdrop.m_stackMax} for level {storedLevel}.");
+                lvlupdrop.m_item = drop.m_item;
+                drops.Add(lvlupdrop);
+            }
+            Physics.SyncTransforms();
+
+            yield break;
+        }
+
         [HarmonyPatch(typeof(TreeBase), nameof(TreeBase.Awake))]
         public static class RandomTreeLevelExtension
         {
@@ -382,31 +414,21 @@ namespace StarLevelSystem.modules
             {
                 if (ValConfig.EnableTreeScaling.Value == false) { return; }
 
-                int storedLevel = __instance.m_nview.GetZDO().GetInt("SLE_Tree", 0);
-                if (storedLevel == 0)
+                int storedLevel = 0;
+                if (ValConfig.UseDeterministicTreeScaling.Value)
                 {
-                    LevelSystem.SelectCreatureBiomeSettings(__instance.gameObject, out string creature_name, out DataObjects.CreatureSpecificSetting creature_settings, out BiomeSpecificSetting biome_settings, out Heightmap.Biome biome);
-                    storedLevel = LevelSystem.DetermineLevel(__instance.gameObject, creature_name, creature_settings, biome_settings, ValConfig.TreeMaxLevel.Value);
-                    __instance.m_nview.GetZDO().Set("SLE_Tree", storedLevel);
-                }
-                if (storedLevel > 1)
-                {
-                    float scale = 1 + (ValConfig.TreeSizeScalePerLevel.Value * storedLevel);
-                    __instance.m_health += (__instance.m_health * 0.1f * storedLevel);
-                    // Logger.LogDebug($"Setting Tree size {scale}.");
-                    __instance.transform.localScale *= scale;
-                    List<DropTable.DropData> drops = new List<DropTable.DropData>();
-                    foreach (var drop in __instance.m_dropWhenDestroyed.m_drops)
+                    storedLevel = CompositeLazyCache.GetOrAddCachedTreeEntry(__instance.m_nview);
+                } else {
+                    storedLevel = __instance.m_nview.GetZDO().GetInt(SLS_TREE, 0);
+                    if (storedLevel == 0)
                     {
-                        DropTable.DropData lvlupdrop = new DropTable.DropData();
-                        // Scale the amount of drops based on level
-                        lvlupdrop.m_stackMin = Mathf.RoundToInt(drop.m_stackMin * (ValConfig.PerLevelLootScale.Value * storedLevel));
-                        lvlupdrop.m_stackMax = Mathf.RoundToInt(drop.m_stackMax * (ValConfig.PerLevelLootScale.Value * storedLevel));
-                        // Logger.LogDebug($"Scaling drop {drop.m_item.name} from {drop.m_stackMin}-{drop.m_stackMax} to {lvlupdrop.m_stackMin}-{lvlupdrop.m_stackMax} for level {storedLevel}.");
-                        lvlupdrop.m_item = drop.m_item;
-                        drops.Add(lvlupdrop);
+                        LevelSystem.SelectCreatureBiomeSettings(__instance.gameObject, out string creature_name, out DataObjects.CreatureSpecificSetting creature_settings, out BiomeSpecificSetting biome_settings, out Heightmap.Biome biome);
+                        storedLevel = LevelSystem.DetermineLevel(__instance.gameObject, creature_name, creature_settings, biome_settings, ValConfig.TreeMaxLevel.Value);
+                        __instance.m_nview.GetZDO().Set(SLS_TREE, storedLevel);
                     }
-                    Physics.SyncTransforms();
+                }
+                if (storedLevel >= 1) {
+                    __instance.StartCoroutine(LevelSystem.ModifyTreeWithLevel(__instance, storedLevel));
                 }
             }
         }
@@ -453,7 +475,15 @@ namespace StarLevelSystem.modules
             [HarmonyPatch(nameof(TreeLog.Awake))]
             [HarmonyPostfix]
             static void SetupAwakeLog(TreeLog __instance) {
-                int level = __instance.m_nview.GetZDO().GetInt("SLE_Tree", 1);
+                int level = 1;
+                if (ValConfig.UseDeterministicTreeScaling.Value) {
+                    level = CompositeLazyCache.GetOrAddCachedTreeEntry(__instance.m_nview);
+                } else {
+                    if (__instance.m_nview.GetZDO() != null) {
+                        //Logger.LogDebug("Checking stored Zvalue for tree level");
+                        level = __instance.m_nview.GetZDO().GetInt(SLS_TREE, 1);
+                    }
+                }
                 UpdateDrops(__instance, level);
                 __instance.m_health += (__instance.m_health * 0.1f * level);
                 __instance.GetComponent<ImpactEffect>()?.m_damages.Modify(1 + (0.1f * level));
@@ -463,6 +493,8 @@ namespace StarLevelSystem.modules
             [HarmonyPostfix]
             internal static void RemoveTreeLogInst(TreeLog __instance) {
                 //Logger.LogDebug("Destroying Treelog");
+                if (__instance.m_nview == null || __instance.m_nview.GetZDO() == null) { return; }
+                CompositeLazyCache.RemoveTreeCacheEntry(__instance.m_nview.GetZDO().m_uid.ID);
                 ZNetScene.instance.Destroy(__instance.gameObject);
             }
 
@@ -476,16 +508,22 @@ namespace StarLevelSystem.modules
                 // pass on the level
                 // Logger.LogDebug($"Getting tree level");
                 int level = 1;
-                if (instance.m_nview.GetZDO() != null) {
-                    //Logger.LogDebug("Checking stored Zvalue for tree level");
-                    level = instance.m_nview.GetZDO().GetInt("SLE_Tree", 1);
+                if (ValConfig.UseDeterministicTreeScaling.Value) {
+                    level = CompositeLazyCache.GetOrAddCachedTreeEntry(instance.m_nview);
+                } else {
+                    if (instance.m_nview.GetZDO() != null)
+                    {
+                        //Logger.LogDebug("Checking stored Zvalue for tree level");
+                        level = instance.m_nview.GetZDO().GetInt(SLS_TREE, 1);
+                    }
                 }
+
                 //Logger.LogDebug($"Got Tree level {level}");
                 UpdateDrops(tchild, level);
                 tchild.m_health += (tchild.m_health * 0.1f * level);
                 go.GetComponent<ImpactEffect>()?.m_damages.Modify(1 + (0.1f * level));
                 //Logger.LogDebug($"Setting tree level {level}");
-                nview.GetZDO().Set("SLE_Tree", level);
+                nview.GetZDO().Set(SLS_TREE, level);
                 // This is the last log point, destroy the parent
                 ZNetScene.instance.Destroy(instance.gameObject);
             }
@@ -500,7 +538,8 @@ namespace StarLevelSystem.modules
                     // Scale the amount of drops based on level
                     lvlupdrop.m_stackMin = Mathf.RoundToInt(drop.m_stackMin * (ValConfig.PerLevelLootScale.Value * level));
                     lvlupdrop.m_stackMax = Mathf.RoundToInt(drop.m_stackMax * (ValConfig.PerLevelLootScale.Value * level));
-                    //Logger.LogDebug($"Scaling drop {drop.m_item.name} from {drop.m_stackMin}-{drop.m_stackMax} to {lvlupdrop.m_stackMin}-{lvlupdrop.m_stackMax} for level {storedLevel}.");
+                    Logger.LogDebug($"Scaling drop {drop.m_item} from {drop.m_stackMin}-{drop.m_stackMax} to {lvlupdrop.m_stackMin}-{lvlupdrop.m_stackMax} for level {level}.");
+                    if (drop.m_item == null) { continue; }
                     lvlupdrop.m_item = drop.m_item;
                     drops.Add(lvlupdrop);
                 }
@@ -784,7 +823,7 @@ namespace StarLevelSystem.modules
         [HarmonyPatch(typeof(SpawnSystem))]
         public static class SpawnSystemSpawnPatch
         {
-            //[HarmonyEmitIL(".dumps")]
+            [HarmonyEmitIL(".dumps")]
             [HarmonyTranspiler]
             [HarmonyPatch(nameof(SpawnSystem.Spawn))]
             static IEnumerable<CodeInstruction> Transpiler(IEnumerable<CodeInstruction> instructions, ILGenerator generator)
@@ -816,6 +855,13 @@ namespace StarLevelSystem.modules
                 .Advance(1)
                 .RemoveInstructions(1)
                 .InsertAndAdvance(new CodeInstruction(OpCodes.Ldc_I4, 0))
+                .MatchStartForward(
+                    new CodeMatch(OpCodes.Callvirt, AccessTools.Method(typeof(ItemDrop), nameof(ItemDrop.SetQuality)))
+                )
+                .RemoveInstructions(1)
+                .InsertAndAdvance(
+                    Transpilers.EmitDelegate(SetItemLevelFish)
+                )
                 .ThrowIfNotMatch("Unable to patch SpawnSystem Spawn set level.");
 
                 return codeMatcher.Instructions();
@@ -827,6 +873,17 @@ namespace StarLevelSystem.modules
                 if (chara == null) { return; }
                 //Logger.LogDebug($"SpawnSystem.Spawn setting without zone control {chara.m_name}");
                 SetCharacterLevelControl(chara, 1);
+            }
+
+            private static void SetItemLevelFish(ItemDrop item, int _providedLevel)
+            {
+                LevelSystem.SelectCreatureBiomeSettings(item.gameObject, out string creature_name, out DataObjects.CreatureSpecificSetting creature_settings, out BiomeSpecificSetting biome_settings, out Heightmap.Biome biome);
+                int determinedLevel = LevelSystem.DetermineLevel(item.gameObject, creature_name, creature_settings, biome_settings, ValConfig.FishMaxLevel.Value);
+                // not sure we need max quality set high
+                item.m_itemData.m_shared.m_maxQuality = 999;
+                item.SetQuality(determinedLevel);
+                item.m_itemData.m_shared.m_scaleByQuality = ValConfig.FishSizeScalePerLevel.Value;
+                item.Save();
             }
 
             private static void SpawnSystemSetCharacterLevelControl(Character chara, int providedLevel)